My assignments

+++ mortgage_calculator.py
@@ -1,5 +1,5 @@
-principal = int(input("Enter the amount you wish to borrow: "))
-annual_interest_rate = int(input("Enter the amount of percentage: "))
+principal = float(input("Enter the amount you wish to borrow: "))
+annual_interest_rate = float(input("Enter the amount of percentage: "))
 monthly_interest = (annual_interest_rate/100)/12
 duration = int(input("Enter the number of duration you wish the loan to extend(in years): "))
 monthly_duration = duration*12
